September 2023 Release.
Auto-restart NodeZero Runners: We've added support for registering NodeZero Runners as a system service using systemd on Linux. This enables auto-restart of the NodeZero Runner when the system reboots or the Runner process fails for any reason.

Subcommand: h3 start-runner-service {runner_name}

Visit the NodeZero Runner documentation for more details.